INC1-185 refactor: change folders structure
